+++ src/lib/components/ui/Currency.svelte
@@ -1,8 +1,15 @@
 <script lang="ts">
-	import type { HTMLImgAttributes } from 'svelte/elements';
+	import type { SvelteHTMLElements } from 'svelte/elements';
 	import {CURRENCIES, type CurrencyName} from '$data/currencies';
+	import AtomIcon from '@components/icons/Atom.svelte';
+	import ElectronIcon from '@components/icons/Electron.svelte';
+	import ProtonIcon from '@components/icons/Proton.svelte';
+	import PhotonIcon from '@components/icons/Photon.svelte';
 
-	interface Props extends HTMLImgAttributes {
+	type SvgProps = SvelteHTMLElements['svg'];
+
+	interface Props extends SvgProps {
+		color?: string;
 		class?: string;
 		icon?: boolean;
 		name: CurrencyName;
@@ -10,12 +17,19 @@
 
 	let { name, icon = true, class: className = '', ...rest }: Props = $props();
 
-
 	const currency = $derived(CURRENCIES[name]);
 </script>
 
 {#if icon}
-	<img src={`/currencies/${currency.icon}.png`} alt={name} class="size-3.5 {className} object-contain inline" {...rest} />
+	{#if currency.icon === 'atom'}
+		<AtomIcon class="inline {className}" color={currency.color} {...rest} />
+	{:else if currency.icon === 'electron'}
+		<ElectronIcon class="inline {className}" color={currency.color} {...rest} />
+	{:else if currency.icon === 'proton'}
+		<ProtonIcon class="inline {className}" color={currency.color} {...rest} />
+	{:else if currency.icon === 'photon'}
+		<PhotonIcon class="inline {className}" color={currency.color} {...rest} />
+	{/if}
 {:else}
 	<span class="currency">{currency.name}</span>
 {/if}