+++ src/lib/components/ui/Value.svelte
@@ -31,7 +31,7 @@
 <span title={currencyObject?.name ?? ''} {...rest}>
 	{prefix}{formatNumber(value, precision)}
 	{#if currency}
-		<Currency name={currency} icon={true} class={currencyClass} />
+		<Currency name={currency} class={currencyClass} />
 	{/if}
 	{postfix}
 </span>